nothing

+++ include/ndapplib/nd_affair.h
@@ -241,7 +241,7 @@ public:
 			m_enable = 0 ;
 			
 			if (m_buf.size() > 0) {
-				for(int i=m_buf.size()-1; i>=0; i--) {
+				for(size_t i=m_buf.size()-1; i>=0; i--) {
 					Undo(m_buf[i].first,m_buf[i].second,m_buf[i].optype) ;
 				}
 			}
